feat(cli): suppression applies before the budget, and on the system prompt
Deliverables 4 (CLI half) and 6 of #712.

**Forget projects into the plane that owns the memory.** `stella memory
forget` wrote a tombstone in `store.db` and nothing else, so recall learned
about it only at the CLI projection layer — after the budget had already
been spent on the memory it was about to discard. It now also marks
`node.superseded_at` in `context.db`, where every candidate reader already
filters. `restore` lifts both, unconditionally: the two writes can disagree
if a forget failed halfway, and a restore that refuses to fix that leaves a
memory no command can bring back.

The `store.db` tombstone stays canonical — surface-aware, carries the
reason, outlives the row — so this is a projection, not a second source of
truth. It is written second on purpose: a failure there leaves the memory
visible with the tombstone recorded, which is recoverable; the opposite
order could hide a memory with no record of why.

**Quarantine cannot be projected, so it is passed in.** It is derived —
a count of untruthful citations, recomputed per read, never stored state —
so there is no row here to tombstone without duplicating a derivation and
letting the copy go stale. `ScopedStore` reads it inside `query`, fresh
per recall, and hands it to `recall_scoped_excluding`. A read failure fails
that leg closed. The CLI's post-recall filter stays as a net.

**The system prompt gets the filter it never had.** Workspace memory files
were pasted into the prefix with no suppression of any kind, so forgetting
one did not stop it shipping. They now go through
`stella_store::SurfaceSuppression` — the same filter every other surface
uses, with the surface's own policy resolved inside it, so an authored file
is suppressed by id and never by resemblance (re-writing a memory by hand
means it). Fail-closed, and the omission is stated in the prompt rather
than left for the model to not notice.

The hostile-fixtures isolation test seeded `store.db` with 16 bytes of
garbage. That still proved the gate never opens it, but it would now also
trip the fail-closed read — making the hostile-memory assertion pass or
fail for a reason unrelated to isolation — so the fixture is a real store.

Three files grew past their ratchet ceilings; the new code moved into
`memory/suppression.rs`, `contextgraph/suppression.rs`, and an `impl Store`
in `forget.rs` instead. `contextgraph.rs`'s tests move to a sibling module
(pure move), which drops it from 1,541 to 715 lines and retires its
baseline entry.

`make gate` green. Part of #712.
